spi: spi-zynqmp-gqspi: Fix erase failure when tx buswidth is 4
Irrespective of the tx buswidth, for erase commands the flash device
expects to receive the address on single line. But when tx buswidth
is set to 4. The driver while issuing erase commands, send the address
on 4 lines. This causes the erase operation to fail. This patch updates
the driver and sends the address, for all erase commands, on single line.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
State: pending