usb: dwc3: Fixed common regulator for multiple controller instance
When the system has multiple usb controller instances - usb0 and usb1,
and has a common regulator for both. Then execution of binding and
unbinding command with the common regulator impacts both controllers
as enable/bind and disable/unbind.

This patch creates a separate regulator based on the controller enable
in DT. The regulator id is based on the USB controller's full name
"node_name@address" of the DT node label.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
State: pending